% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_measures.R
\name{calculate_confidence_intervals}
\alias{calculate_confidence_intervals}
\title{Calculate confidence intervals from a boot_strap object}
\usage{
calculate_confidence_intervals(
  boot_strap,
  quantiles = c(lower95 = 0.025, lower80 = 0.1, median = 0.5, upper80 = 0.9, upper95 =
    0.975),
  type = "direct"
)
}
\arguments{
\item{boot_strap}{A boot_strap object}

\item{quantiles}{A named vector of quantile bounds, each between 0 and 1}

\item{type}{One of either \code{"direct"}, \code{"relative"}, or \code{"both"}. Use
\code{"direct"} to calculate confidence intervals around all measured KPIs
for each group separately, use \code{"relative"} to calculate the confidence
intervals around the ratios of pair-wise combinations of all measured KPIs
for all groups, and \code{"both"} to calculate both.}
}
\value{
Returns a data frame with quantiles for all groups/comparisons
defined in \code{type}, for all measures in the boot_strap object.
}
\description{
Calculate confidence intervals from a boot_strap object
}
