% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_measures.R
\name{/.boot_strap}
\alias{/.boot_strap}
\title{Division function for boot_strap objects}
\usage{
\method{/}{boot_strap}(numerator, denominator)
}
\arguments{
\item{numerator}{A boot_strap object as the numerator of the division}

\item{denominator}{A boot_strap object or a numeric as the denominator}
}
\description{
Division function for boot_strap objects
}
